requirements.txt
# Crypto Perps Tracker - Python Dependencies
# Core dependencies for fetching and analyzing perpetual futures data

# HTTP requests
requests>=2.31.0

# Data validation (new architecture)
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Configuration (required for Discord integration)
pyyaml>=6.0

# Database (Phase 3+)
sqlalchemy>=2.0.0

# Visualization (required for chart generation)
matplotlib>=3.7.0
mplcyberpunk>=0.7.6
plotly>=5.17.0

# Dashboard (Phase 4)
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# Websockets (Phase 3)
websocket-client>=1.6.0

# Machine Learning (Advanced Features)
scikit-learn>=1.3.0
scipy>=1.11.0

# Message Queue (Advanced Features)
redis>=5.0.0

# Development dependencies (optional)
# pytest>=7.3.0
# black>=23.3.0
# flake8>=6.0.0
